Extract the Session login and logout lifecycle modules
Login claim, logout finalize and disconnect cleanup were embedded in
`session/mod.rs`. Move the exact current behaviour into a private
`session/lifecycle`, concrete database calls included.

  - `lifecycle/login.rs` — the single-live-session character login claim.
  - `lifecycle/logout.rs` — `complete_logout`, the disconnect save and the
    `player_logout` guard the save path sets.
  - `lifecycle/cleanup.rs` — registry, visibility, canonical map and object
    accessor teardown.

The orderings this code must preserve are documented beside the code that owns
them rather than in a note that can drift:

  - cleanup tears down publication before ownership; a session still published
    while its map entry is gone is exactly the window that produces stale
    broadcasts;
  - the disconnect save settles loot first, then the character, then the
    account-wide collections, then the three offline marks;
  - `complete_logout` keeps the represented player alive, because C++
    `WorldSession::LogoutPlayer(true)` saves while `_player` still exists and
    clearing it early would make the later save a no-op.

Nothing is redesigned. The concrete DB calls stay exactly as they were behind
this one private seam so #200 can replace a single boundary after #187 freezes
the focused Player contract, instead of searching the whole Session. The
packet handlers keep living under `handlers/` where the handler contract
requires them.

Verified as a pure move: the function multiset across `session/mod.rs` plus
the new modules is identical before and after (2609 both sides). The only
surface change is `complete_logout` private -> `pub(in crate::session)`, which
the Session driver's logout timer needs and which does not leave the family.

Tests (`session/tests/lifecycle.rs`, 7 cases, all through the real entry
points): claim exclusivity and idempotent re-claim by the owner; releasing
without a claim not freeing another session's; timed logout emitting
`SMSG_LOGOUT_COMPLETE`, disconnecting, and keeping the player for the save;
cleanup releasing the claim so a disconnect cannot leave a character
unloggable; cleanup with no selected character; the disconnect-save short path
without a player; and the logout flag round trip.

Inventories: `complete_logout` re-recorded with its new visibility, three
`unregister_from_player_registry` registry rows re-pathed to
`crate::session::lifecycle::cleanup` with identical fingerprints and the total
unchanged at 570, session hotspot set to measured rows, and the pinned
`#[path]` count 20 -> 21 for the new test mount.

Refs #133, #183, #187, #200. Closes #184.
