+++ tools/architecture/runtime-ownership-ledger.json
@@ -1870,9 +1870,9 @@
       "entries": [
         {
           "path": "crates/wow-world/src/session/mod.rs",
-          "production_lines": 75010,
-          "test_lines": 95123,
-          "total_lines": 170133,
+          "production_lines": 75081,
+          "test_lines": 95247,
+          "total_lines": 170328,
           "owner": "wow-world WorldSession/application monolith.",
           "open_retirement_issues": [
             140,
